+++ cx-portal-shared-components/src/components/basic/ImageGallery/index.tsx
@@ -18,53 +18,26 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
+import { ImageItem } from './ImageItem'
 import { ImageType } from './types'
-import { Typography } from '../Typography'
 
-export const ImageGallery = ({ gallery }: { gallery: ImageType[] }) => (
+export const ImageGallery = ({
+  gallery,
+  grid = false,
+}: {
+  gallery: ImageType[]
+  grid?: boolean
+}) => (
   <div
     style={{
-      display: 'flex',
-      flexWrap: 'wrap',
-      marginRight: '-15px',
-      marginLeft: '-15px',
-      boxSizing: 'border-box',
+      display: grid ? 'grid' : 'flex',
+      gap: '30px',
+      placeContent: 'center',
+      gridTemplateColumns: 'repeat(3, 1fr)',
     }}
   >
     {gallery.map((image, index) => (
-      <div
-        style={{
-          position: 'relative',
-          width: '100%',
-          paddingRight: '15px',
-          paddingLeft: '15px',
-          flex: '0 0 33.333333%',
-          maxWidth: '33.333333%',
-          boxSizing: 'border-box',
-        }}
-        key={index}
-      >
-        <img
-          src={image.url}
-          alt={image.text}
-          style={{
-            width: '100%',
-            borderRadius: '20px',
-            height: '250px',
-            objectFit: 'cover',
-          }}
-        />
-        <Typography
-          variant="body2"
-          sx={{
-            color: '#888888',
-            margin: '5px 0',
-            fontSize: '14px',
-          }}
-        >
-          {image.text}
-        </Typography>
-      </div>
+      <ImageItem key={index} image={image} />
     ))}
   </div>
 )