+++ cx-portal-shared-components/src/components/basic/ImageGallery/types.ts
@@ -19,6 +19,16 @@
  ********************************************************************************/
 
 export type ImageType = {
-  url: string
-  text: string
+  url?: string
+  text?: string
+  size?:
+    | 'small-rectangle'
+    | 'small-square'
+    | 'medium-rectangle'
+    | 'medium-square'
+    | 'large-rectangle'
+    | 'large-square'
+  hover?: boolean
+  borderRadius?: boolean
+  shadow?: boolean
 }